+++ lib/src/NavigationMock/Components/BottomTabs.tsx
@@ -1,19 +1,14 @@
 import React, { Component } from 'react';
-import { View } from 'react-native';
-import { LayoutComponent } from './LayoutComponent';
-import { ComponentProps } from './ComponentProps';
-import { connect } from './connect';
+import { LayoutComponent } from '../LayoutComponent';
+import { ComponentProps } from '../ComponentProps';
+import { connect } from '../connect';
 
 export const BottomTabs = connect(
   class extends Component<ComponentProps> {
-    renderScreens() {
+    render() {
       return this.props.layoutNode.children.map((child) => {
         return <LayoutComponent key={child.nodeId} layoutNode={child} />;
       });
-    }
-
-    render() {
-      return <View>{this.renderScreens()}</View>;
     }
   }
 );
