+++ lib/src/NavigationMock/Components/ComponentScreen.tsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { Button, View, Text } from 'react-native';
-import { Navigation } from 'react-native-navigation';
-import { ComponentProps } from './ComponentProps';
-import { VISIBLE_SCREEN_TEST_ID } from './constants';
-import { LayoutStore } from './LayoutStore';
-import { connect } from './connect';
+import { Navigation } from '../..';
+import { ComponentProps } from '../ComponentProps';
+import { VISIBLE_SCREEN_TEST_ID } from '../constants';
+import { LayoutStore } from '../LayoutStore';
+import { connect } from '../connect';
 import { TopBar } from './TopBar';
 
 export const ComponentScreen = connect(
